Updated libft/math/ft_pfactorial.c file

+++ libft/math/ft_pfactorial.c
@@ -6,13 +6,13 @@
 /*   By: alkuzin <[null]@student.42.fr>             +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/29 13:11:52 by alkuzin           #+#    #+#             */
-/*   Updated: 2023/11/29 13:15:23 by alkuzin          ###   ########.fr       */
+/*   Updated: 2023/12/31 15:32:51 by alkuzin          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ft_math.h"
 
-unsigned int ft_pfactorial(unsigned int n, unsigned int m)
+unsigned long ft_pfactorial(unsigned int n, unsigned int m)
 {
     unsigned long result;
